mental-health-responder/requirements.txt
# requirements.txt
streamlit
python-dotenv
requests
transformers
torch  # Required for transformers